add bookType

+++ modal/role.go
@@ -1,7 +1,7 @@
 package modal
 
 type Role struct {
-	Id          uint16 `json:"id,omitempty" form:"id"`
+	Id          uint   `json:"id,omitempty" form:"id"`
 	Role_name   string `json:"role_name,omitempty" form:"role_name"`
 	Role_weight uint16 `json:"role_weight,omitempty" form:"role_weight"`
 }
@@ -15,7 +15,7 @@ func NewRole() *Role {
 }
 
 type RoleListResult struct {
-	Id          uint16 `json:"id,omitempty" form:"id"`
+	Id          uint   `json:"id,omitempty" form:"id"`
 	Role_name   string `json:"role_name,omitempty" form:"role_name"`
 	Role_weight string `json:"role_weight,omitempty" form:"role_weight"`
 }
